refactor(payroll): lot 6b — simulateur cotisations extrait en Action (SimulateCotisations), contrôleur allégé
