+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - **feat(ai): A5 — audit conversation→action→effet : journal d'exécution d'outils `ai_tool_executions` avec marqueur « via assistant IA » (Closes #6852, BC-23, EPIC #6846).** Chaque étape de la chaîne est tracée de façon rejouable : (1) outil de LECTURE exécuté pendant le chat → ligne `stage=executed` reliée à la conversation ; (2) écriture sensible → ligne de proposition `stage=confirmation_required` avec `pending_action_id` ; (3) confirmation humaine → ligne d'exécution reliée au MÊME `pending_action_id` (la proposition porte la `conversation_id`, la chaîne se reconstitue par jointure) ; (4) refus humain → `stage=rejected`, aucun effet métier. Colonnes : `tool_name`, `tool_input` (JSON SANITISÉ — PII masquées via PrivacySanitizer, A6 #6853), `stage`, `success`, `result_summary`, `source='assistant'`, horodatage serveur ; index (company_id, created_at), conversation_id, pending_action_id. Câblage : `AIAuditLogger::logToolExecution()` (nouveau), boucle d'outils de l'`Orchestrator` (chat), `AIGatewayController::confirmAction`/`rejectAction`. Migration tenant idempotente + miroir fixture `CreatesMvpSchema` (garde parity #5443). Tests Feature `AIToolExecutionAuditTest` : lecture journalisée avec conversation, chaîne écriture proposition→exécution par pending_action_id, refus sans effet. Vérifié : 3 tests verts (PHP 8.4 + PG16), Pint OK, php -l OK. Aucun changement de contrat API (routes/verbatim inchangés, écritures additives).
 - **fix(tests): AbsenceTypesTenantUniqueTest — rejet du doublon intra-tenant isolé dans une transaction imbriquée, fin du 25P02 (Part of #6954, volet restant après #6955).** Le test `test_duplicate_code_within_same_tenant_is_rejected` assertait la violation 23505 attendue de l'index unique `(company_id, code)` (#5967) avec un `expectException` hors transaction : la violation laissait la transaction du test (posée par RefreshDatabase) en état « aborted » et le `SET search_path` de restauration (tearDown/withinTenant) échouait en `SQLSTATE[25P02]` → le fichier restait rouge sur main. Correctif : pattern repo déjà utilisé par les tests Travel — `expectException` dans le contexte tenant + création fautive dans un `DB::transaction` imbriqué (savepoint rollbacké, l'état aborted est guéri avant la rethrow). Aucun changement de comportement applicatif (le 23505 est toujours levé et asserté). Vérifié : 4 tests / 19 assertions verts en local (PHP 8.4 + PG16), Pint OK, php -l OK.
 - **feat(ai): B3a — outil écriture `absence_decision` (approbation / refus motivé d'une demande d'absence) déclaré au contrat A3, exécution après confirmation (Closes #6856, BC-06 LEAVE, EPIC #6846).** Premier outil `write` du programme B hors legacy #5625 : `AbsenceDecisionToolCatalog` (module façade Absence, BC-06) enregistré au boot par `AbsenceServiceProvider` (garde d'idempotence #6947) ; le tool_call ne crée qu'une `PendingAction` (flux A4) — **aucun effet de bord avant confirmation** (`POST /ai/actions/{id}/confirm|reject`). L'exécution passe par les Actions canoniques Planning `ApproveAbsence`/`RejectAbsence` → `AbsenceService` (PA2-ARCH-002, module propriétaire du domaine absence/congé) : mêmes transitions de statut, mêmes événements métier (`AbsenceApproved`/`AbsenceRejected`) et même journal d'audit que les endpoints REST `POST|PUT /api/v1/absences/{absence}/approve|reject`. Câblage complet : `config/ai.php` (`write_tools` + matrice `tool_permissions` : rôle manager, permission `absences.approve`), entrée `ai_tool_registry` (AIToolRegistrySeeder), handler `WriteActionRunner::decideAbsence` injectant les deux Actions (parité RBAC REST : manager du tenant uniquement, isolation tenant fail-closed, motif de refus obligatoire ≤ 1000 caractères, mapping propre `ABSENCE_NOT_PENDING` / `INSUFFICIENT_LEAVE_BALANCE`), résumé de confirmation FR dans `IntentEngine` (approbation vs refus motivé), arête MAT-002 `BC-06 → BC-23` documentée. Tests Feature `AbsenceDecisionToolTest` : proposition → confirmation → exécution → événement, refus sans motif refusé, RBAC (employé refusé avant et à la confirmation), isolation cross-tenant, décision invalide, absence déjà décidée, déduction de solde sur type déductible. Vérifié : Pint OK, PHPStan strict (niveau 8) 0 erreur sur le périmètre. Aucun changement de comportement des outils existants (`create_absence`, `approve_absence` conservés — additif).
+- **refactor(payroll): lot 6b — simulateur cotisations/IR extrait en Action `SimulateCotisations`.** CotisationSimulationController allégé (−120 l.) : résolution des règles tenant+période effective (#1924, échecs audités), garde placeholder #1872 via `PayrollPlaceholderAcknowledgementRequiredException` (422, même payload), ventilation salarial/patronal #2220 et calcul via le pipeline unique des bulletins conservés à l'"'"'identique ; l'"'"'interface garde autorisation (manager), validation, corrélation #1874, audit HTTP de l'"'"'acceptation placeholder et présentation du contrat (#1869). Vérifié : php -l OK, Pint 1.30.5 PASS, aucune migration.
 - **feat(crm): répertoire tenant — endpoints de liste paginés leads/accounts/contacts/opportunities livrés (Closes #6977, BC-11 CRM, contrat #5712).** Le dashboard client web livré par #5715 (`/crm/leads|accounts|contacts|pipeline`) appelait 4 endpoints inexistants → sections mortes (404) dès qu'un tenant active le flag `crm` (UI fantôme). Livraison du contrat : `CrmDirectoryController` (GET `/crm/{leads,accounts,contacts,opportunities}`, pagination `per_page`≤100 + meta, lecture = managers du tenant via middleware `api.manager`, isolation tenant `company_id` fail-closed) ; exclusions : leads convertis (`converted_at`), comptes/contacts archivés ; contacts avec compte parent minimal embarqué (scopé tenant). 3 Resources (`CrmAccountResource`, `CrmContactResource`, `CrmOpportunityResource`) ; OAS : 4 chemins + schémas `CrmAccount`/`CrmContact`. Test Feature `CrmDirectoryListTest` (isolation tenant, pagination, exclusions, account embarqué, 403 employé). Aucun changement de comportement existant (routes additives). Vérifié : Pint OK, PHPStan strict 0, YAML OAS valide (731 paths).
 
 - **fix(ops): login super-admin (API + web) — exceptions loggées (classe + message) au lieu d'un 500 muet (Part of #6974).** Sur DEV, `POST /api/v1/platform/auth/login` renvoie un 500 inexpliqué même APRÈS la garde #6956 (hash non NULL) — la cause réelle reste à identifier côté logs. Les deux contrôleurs (`Core/Auth/.../PlatformAuthController` API et `Http/Controllers/Web/PlatformAuthController`) enveloppent désormais leur login : toute exception est loggée en `structured` (`platform.login.unexpected_error[.web]`, email + `exception` + message) et le client reçoit la réponse d'erreur générique habituelle (aucun détail exposé, aucun changement de comportement nominal). Vérifié : garde governance/scénarios alignée, aucun changement de contrat API.
